Reorganize code into Compressor/Decompressor classes.
