+++ lib/context2.js
@@ -15,9 +15,10 @@ import {VocabTermDecoder} from './codecs/VocabTermDecoder.js';
  * @returns {Promise<object>} - The active context instance.
  */
 export async function getActiveContext({
-  contextStack, contextMap, termToId, idToTerm, obj, documentLoader,
-  scoped = false, encodedMap
+  obj, encodedMap, contextStack, transformer, scoped = false
 }) {
+  const {contextMap, termToId, idToTerm, documentLoader} = transformer;
+
   // FIXME: change above params to include `processorState` or better named
   // structure
   const stackTop = contextStack.length;
